Modified some code to reduce warnings in setup.rb (e.g. removed unused variables) (untouched matrix_source.c and vector_source.c to  avoid conflict since  these are now extensively developed)
git-svn-id: svn+ssh://rubyforge.org/var/svn/rb-gsl/trunk/rb-gsl@69 6e764f74-f39f-46f8-8c54-8307d62afe8d

+++ ext/interp.c
@@ -189,11 +189,11 @@ static VALUE rb_gsl_interp_evaluate(VALUE obj, VALUE xxa, VALUE yya, VALUE xx,
   Data_Get_Struct(obj, rb_gsl_interp, rgi);
   ptrx = get_vector_ptr(xxa, &stridex, &size);
   if (size != rgi->p->size ){
-    rb_raise(rb_eTypeError, "size mismatch (xa:%d != %d)", size, rgi->p->size);
+    rb_raise(rb_eTypeError, "size mismatch (xa:%d != %d)",  (int) size, (int) rgi->p->size);
   }
   ptry = get_vector_ptr(yya, &stridey, &size);
   if (size != rgi->p->size ){
-    rb_raise(rb_eTypeError, "size mismatch (ya:%d != %d)", size, rgi->p->size);
+    rb_raise(rb_eTypeError, "size mismatch (ya:%d != %d)", (int) size, (int) rgi->p->size);
   }
   if (CLASS_OF(xx) == rb_cRange) xx = rb_gsl_range2ary(xx);
   switch (TYPE(xx)) {
